Extract challenge verification logic to service
